fix(trading): extract values from market candles
