+++ libs/markets/src/lib/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './use-market-oracle';
 export * from './use-oracle-markets';
 export * from './use-oracle-proofs';
 export * from './use-oracle-spec-binding-data';
+export * from './use-candles';
